Remove thread pool executor
Usage of the thread pool was a temporary hack to run requests in
parallel, but it was highly inefficient since it created event loops for
every request and did not use asynchronous sockets. Asyncio streams are
now used to use a single thread efficiently. This also lays the
groundwork for distributed (multiprocess) serving.